Add files via upload
These are a few of some of my favorite AP Computer Science projects, ranging from a simple calculator that can add, subtract, multiply, and divide to a Name Generator that generates the Marvel superhero or villain that you are based on your inputs. There's also a fun guessing game where you can guess the secret number, and see how close you are to the number. But if you're looking to find something on the more serious side of things there's conditional challenges that test different skills, and a fun use of 99 bottles of "milk" to practice the use of loops and string concatenation.